+++ packages/utils/interactive-hooks/src/useAsPressable.ts
@@ -12,6 +12,7 @@ import { usePressability } from './usePressability';
 /**
  * hover specific state and callback helper
  */
+// eslint-disable-next-line @typescript-eslint/ban-types
 function useHoverHelper<T extends object>(props: IWithPressableOptions<T>): [IWithPressableOptions<object>, IHoverState] {
   const [hoverState, setHoverState] = React.useState({ hovered: false });
   const onHoverIn = React.useCallback(
@@ -39,6 +40,7 @@ function useHoverHelper<T extends object>(props: IWithPressableOptions<T>): [IWi
 /**
  * focus specific state and callback helper
  */
+// eslint-disable-next-line @typescript-eslint/ban-types
 function useFocusHelper<T extends object>(props: IWithPressableOptions<T>): [IWithPressableOptions<object>, IFocusState] {
   const [focusState, setFocusState] = React.useState({ focused: false });
   const onFocus = React.useCallback(
@@ -66,6 +68,7 @@ function useFocusHelper<T extends object>(props: IWithPressableOptions<T>): [IWi
 /**
  * press specific state and callback helper
  */
+// eslint-disable-next-line @typescript-eslint/ban-types
 function usePressHelper<T extends object>(props: IWithPressableOptions<T>): [IWithPressableOptions<object>, IPressState] {
   const [pressState, setPressState] = React.useState({ pressed: false });
 
@@ -98,6 +101,7 @@ function usePressHelper<T extends object>(props: IWithPressableOptions<T>): [IWi
  * The useAsPressable hook adds a simple state change function for listening to hover, press, and focus events on the base pressability implementation
  * @param props - input props for the component, mixed in with pressable and pressability options
  */
+// eslint-disable-next-line @typescript-eslint/ban-types
 export function useAsPressable<T extends object>(props: IWithPressableOptions<T>): IPressableHooks<T> {
   const [hoverProps, hoverState] = useHoverHelper(props);
   const [focusProps, focusState] = useFocusHelper(props);
@@ -115,6 +119,7 @@ export function useAsPressable<T extends object>(props: IWithPressableOptions<T>
  * as each of these calls will create a new instance of the Pressability class.
  * @param props - input props for the component
  */
+// eslint-disable-next-line @typescript-eslint/ban-types
 export function useFocusState<T extends object>(props: IWithPressableOptions<T>): [IWithPressableEvents<T>, IFocusState] {
   const [focusProps, focusState] = useFocusHelper(props);
   return [{ ...props, ...usePressability({ ...props, ...focusProps }) }, focusState];
@@ -125,6 +130,7 @@ export function useFocusState<T extends object>(props: IWithPressableOptions<T>)
  * as each of these calls will create a new instance of the Pressability class.
  * @param props - input props for the component
  */
+// eslint-disable-next-line @typescript-eslint/ban-types
 export function usePressState<T extends object>(props: IWithPressableOptions<T>): [IWithPressableEvents<T>, IPressState] {
   const [pressProps, pressState] = usePressHelper(props);
   return [{ ...props, ...usePressability({ ...props, ...pressProps }) }, pressState];
@@ -135,6 +141,7 @@ export function usePressState<T extends object>(props: IWithPressableOptions<T>)
  * as each of these calls will create a new instance of the Pressability class.
  * @param props - input props for the component
  */
+// eslint-disable-next-line @typescript-eslint/ban-types
 export function useHoverState<T extends object>(props: IWithPressableOptions<T>): [IWithPressableEvents<T>, IHoverState] {
   const [hoverProps, hoverState] = useHoverHelper(props);
   return [{ ...props, ...usePressability({ ...props, ...hoverProps }) }, hoverState];